# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Intern Management API",
    "version": "1.0.0",
    "description": "API quản lý thực tập sinh - WordPress Plugin"
  },
  "servers": [
    {
      "url": "/wp-json/intern/v1",
      "description": "WordPress REST API"
    }
  ],
  "components": {
    "securitySchemes": {
      "jwtAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Nhập JWT token của bạn vào đây (không cần thêm chữ Bearer ở trước)"
      }
    },
...
  },
  "security": [
    {
      "wpNonce": []
    }
  ],
  "paths": {
    "/projects": {
      "get": {
        "summary": "Lấy danh sách dự án",
        "tags": ["Project"],
        "security": [
          {
            "wpNonce": []
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách dự án",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "items": [
                      {
                        "id": "5",
                        "name": "adasd",
                        "description": "asd",
                        "status": "waiting",
                        "manager_id": "1",
                        "start_date": "2026-02-03",
                        "end_date": "2026-02-05",
                        "created_at": "2026-02-28 10:53:26",
                        "updated_at": "2026-02-28 10:53:26",
                        "deleted_at": null
                      },
                      {
                        "id": "4",
                        "name": "asdasdasd",
                        "description": "asdasdasd",
                        "status": "in_progress",
                        "manager_id": "1",
                        "start_date": "2026-02-03",
                        "end_date": "2026-02-17",
                        "created_at": "2026-02-23 23:57:30",
                        "updated_at": "2026-02-23 23:57:30",
                        "deleted_at": null
                      },
                      {
                        "id": "3",
                        "name": "asdasd11",
                        "description": "asdasdasd11",
                        "status": "waiting",
                        "manager_id": "1",
                        "start_date": "2026-02-09",
                        "end_date": "2026-02-25",
                        "created_at": "2026-02-23 22:50:38",
                        "updated_at": "2026-02-23 22:50:48",
                        "deleted_at": null
                      }
                    ],
                    "meta": {
                      "current_page": 1,
                      "per_page": 10,
                      "total": 3,
                      "last_page": 1
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Tạo dự án",
        "tags": ["Project"],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "description",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "status",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["in_progress", "waiting", "on_hold", "completed"],
              "example": "in_progress"
            }
          },
          {
            "name": "manager_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "name": "start_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content" : {
              "application/json": {

              }
            }
          }
        }
      }
    },
    "/projects?page=page&per_page=per_page": {
      "get": {
        "summary": "Lấy danh sách dự án",
        "tags": ["Project"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Trang hiện tại"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Số lượng mỗi trang"
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách dự án",
            "content": {
              "application/json": {
              }
            }
          }
        }
      }
    },
...
  }
}